LayoutTests/imported/w3c/web-platform-tests/cookie-store/cookieStore_special_names.https.any.serviceworker-expected.txt


FAIL cookieStore.set with __Secure- name on secure origin promise_test: Unhandled rejection with value: object "SecurityError: The operation is insecure."
FAIL cookieStore.set of expired __Secure- cookie name on secure origin promise_test: Unhandled rejection with value: object "SecurityError: The operation is insecure."
FAIL cookieStore.delete with __Secure- name on secure origin promise_test: Unhandled rejection with value: object "SecurityError: The operation is insecure."
FAIL cookieStore.set with __Host- name on secure origin promise_test: Unhandled rejection with value: object "SecurityError: The operation is insecure."
FAIL cookieStore.set of expired __Host- cookie name on secure origin promise_test: Unhandled rejection with value: object "SecurityError: The operation is insecure."
FAIL cookieStore.delete with __Host- name on secure origin promise_test: Unhandled rejection with value: object "SecurityError: The operation is insecure."
FAIL cookieStore.set with __Host- prefix and a domain option promise_rejects_js: function "function() { throw e }" threw object "SecurityError: The operation is insecure." ("SecurityError") expected instance of function "function TypeError() {
    [native code]
}" ("TypeError")
FAIL cookieStore.set with __Host- prefix a path option promise_test: Unhandled rejection with value: object "SecurityError: The operation is insecure."
FAIL cookieStore.set with malformed name. assert_equals: cookieStore thrown an incorrect exception - expected "TypeError" but got "SecurityError"
